l2w: long tap count work added
This way the Wake/Sleep event is triggered while the
finger is still on the touchscreen in the Logo area.
If it's left or finger released, the async work is breaking
before triggering the power on/off event.
Much more convenient than to wait for the finger
release event, user will actually know if he tapped the
wrong place, before releasing finger, and can move it a
bit lower/higher and the event will be triggered while
still touching the screen.